minisat: support ipasir_solve_final
To support simplification, the glue class has to be inherited from
the SimpSolver class. Therefore, we also need to patch this file,
and include the header during compilation.

The implementation of the ipasir_solve_final method checks whether
the number of calls to solve is 0, and only in this case
simplification is used before solve is actually called.