Document multithreading assumptions of ipasir_set_learn
Whithout this assumption, a solver could call ipasir_set_learn from
a different thread than the thread from which the caller invoked
ipasir_solve.  It could even call the callback concurrently from
different threads.  Given that callbacks are likely unprepared for
the challenges of concurrency and that existing solvers already
seem to fulfil the assumption, it seem reasonable to add it to the
interface specification.